+++ sdk/go/impart/getConnector.go
@@ -74,14 +74,20 @@ type GetConnectorResult struct {
 
 func GetConnectorOutput(ctx *pulumi.Context, args GetConnectorOutputArgs, opts ...pulumi.InvokeOption) GetConnectorResultOutput {
 	return pulumi.ToOutputWithContext(context.Background(), args).
-		ApplyT(func(v interface{}) (GetConnectorResult, error) {
+		ApplyT(func(v interface{}) (GetConnectorResultOutput, error) {
 			args := v.(GetConnectorArgs)
-			r, err := GetConnector(ctx, &args, opts...)
-			var s GetConnectorResult
-			if r != nil {
-				s = *r
+			opts = internal.PkgInvokeDefaultOpts(opts)
+			var rv GetConnectorResult
+			secret, err := ctx.InvokePackageRaw("impart:index/getConnector:GetConnector", args, &rv, "", opts...)
+			if err != nil {
+				return GetConnectorResultOutput{}, err
 			}
-			return s, err
+
+			output := pulumi.ToOutput(rv).(GetConnectorResultOutput)
+			if secret {
+				return pulumi.ToSecret(output).(GetConnectorResultOutput), nil
+			}
+			return output, nil
 		}).(GetConnectorResultOutput)
 }