+++ sdk/nodejs/getConnector.ts
@@ -19,7 +19,6 @@ import * as utilities from "./utilities";
  * ```
  */
 export function getConnector(args: GetConnectorArgs, opts?: pulumi.InvokeOptions): Promise<GetConnectorResult> {
-
     opts = pulumi.mergeOptions(utilities.resourceOptsDefaults(), opts || {});
     return pulumi.runtime.invoke("impart:index/getConnector:GetConnector", {
         "connectorTypeId": args.connectorTypeId,
@@ -87,7 +86,13 @@ export interface GetConnectorResult {
  * ```
  */
 export function getConnectorOutput(args: GetConnectorOutputArgs, opts?: pulumi.InvokeOptions): pulumi.Output<GetConnectorResult> {
-    return pulumi.output(args).apply((a: any) => getConnector(a, opts))
+    opts = pulumi.mergeOptions(utilities.resourceOptsDefaults(), opts || {});
+    return pulumi.runtime.invokeOutput("impart:index/getConnector:GetConnector", {
+        "connectorTypeId": args.connectorTypeId,
+        "id": args.id,
+        "isConnected": args.isConnected,
+        "name": args.name,
+    }, opts);
 }
 
 /**