+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [0.8.0] - 2024-11-11
+
+### Added
+
+- Label resource
+- TagMetadata resource
+- labels attribute to rule scripts, rule test cases, lists, monitors, tags
+- description attribute to lists and to test cases messages and assertions
+
+### Changed
+- List implementation to use PUT and PATCH items endpoint
+
 ## [0.7.0] - 2024-09-23
 
 ### Added